Added traffic funcs and traffic_compressor binary
Lots of changes:
  - Log now includes line and file
  - khash library (from klib)
  - Traffic compressor which accepts a traffic file in the old format
    and generates compatible trace files.
  - traffic.{c,h} for working with traffic files

+++ include/dataplane.h
@@ -11,7 +11,7 @@ typedef uint32_t link_id_t;
 typedef uint32_t pair_id_t;
 
 /* bandwidth type variable */
-typedef double bw_t;
+typedef float bw_t;
 
 struct link_t {
   link_id_t id;
